fix warnings from puma
Puma had a noisy warning about header keys being invalid, and there was
another warning about Puma booting up when capybara first starts.

I also disabled the `Style/StringHashKeys` rule, as it wants to change
this. It's not a very safe rule for auto-correct, so probably best to
leave off.