introduce CallableJob
`CallableJob` is a generic job that allows us to easily background any
class or constant that implements a `call` method without needing to
introduce a new type of job. This also makes zero downtime deploys
easier, since we don't need an extra step to introduce the new job
before adding any code that references it, though we still need to make
sure the callable is in production before adding any code that will
depend on it via `CallableJob`. This is less likely to be a problem,
however, since often times the logic that we want to background is
already in production and is slowing down the request lifecycle. We can
quickly change the logic to background it like so:

```diff
- MyThing.call(args)
+ CallableJob.perform_later("MyThing", args)
```