remove VCR
I've been doing some research on the best way to test against APIs and
my thinking is now that VCR is maybe not the best option. Having worked
some with VCR in other scenarios, I've found it to be a lot of work to
maintain, scrubbing data from VCR cassettes and working around changes
to the data. Instead, I'm now thinking this is a somewhat better
approach to testing against 3rd party APIs:

1) Test against the real API in dev mode first.
2) Once it is working with the real API, enable Webmock (and turn off
  your wifi if your Capybara tests visit external sites in the browser).
  Write a test adapter to fake out the API you're testing against and
  inject that as a replacement for the real integration. This should not
  require changing the actual tests, but should result in the unchanged
  tests passing.
3) Allow toggling between the fake and real implementation, probably via
  some env variable (`FAKE_APIS=false rspec spec`). The tests should run
  the same against the real thing and the fake. On CI we will run
  against the fake version, with maybe a daily build against the real
  thing.

This approach gives a more programmatic way of managing the integration.
The data in the fakes is not as realistic as would be provided by VCR,
but our test APIs can put the focus on what data is most relevant, and
making sure the tests pass against both the real implementation and the
fake should help ensure that they're still accurate. We can potentially
use [`capybara-discoball`][2] to spin up local fake APIs to stand in for
the 3rd party.

More thoughts in [this thoughtbot video][2].

[1]: https://github.com/thoughtbot/capybara_discoball
[2]: https://thoughtbot.com/upcase/videos/testing-interaction-with-3rd-party-apis